add fourier based AHRS position estimator
Description :

monocopter gyropter support, the AHRS pitch/roll/yaw are determined with the help of a Fourier transform.

This effectively adds support for the gyropter/ mono blade drones

Without that, when the frame is continuously rotating, the IMU with the standard DCM and kalman filter get dizzy and diverge.
With this new AHRS algorithm, the pitch/roll/yaw are no longer the instantaneous frame angle but the solid disk rotor one.
The solid disk rotor is the area swept by the rotation of the frame.